requirements.txt
Django==4.2
django-cors-headers==3.10.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Pillow==9.0.1
psycopg2-binary==2.9.9
django-phonenumber-field==7.2.0
phonenumbers==8.13.27
requests==2.31.0
drf-writable-nested==0.7.0
